+++ HISTORY.md
@@ -31,6 +31,9 @@
 * **Translation Updates:**
     * German
 
+### 2022-04-04
+* objects: added speech-balloon-value-association support for ASK menu items
+
 ### 2022-04-03
 * objects: added shortcut support for ASK menu items